Create Big Data Technologies
The explosion of data has made it increasingly important to be able to analyze vast amounts of data in real-time. Big data technologies, such as Spark and Hadoop, have made it possible to perform complex analyses on large datasets, and to do so in a fraction of the time it would have taken using traditional methods.